Refactored season adapter and interfaces
