+++ src/databaseAdapters/databaseInterfaces/index.ts
@@ -1,5 +1,7 @@
 import IUserDbAdapter from './user';
+import ISeasonDbAdapter from './season';
 
 export default interface IDatabaseAdapter {
   user: IUserDbAdapter;
+  season: ISeasonDbAdapter;
 }
